+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
+## [1.10.0-rc.0] - 2019-08-16
+### Added
+- `postOperationsToSpecificEndpoint` in to api-caller
+
 ## [1.9.0] - 2019-08-12
 
 ## [1.9.0-rc.2] - 2019-07-26
@@ -196,7 +200,8 @@ for further information about branching and tagging conventions.
 
 ## [1.3.0] - 2019-03-01
 
-[Unreleased]: https://github.com/tokend/new-js-sdk/compare/1.9.0...HEAD
+[Unreleased]: https://github.com/tokend/new-js-sdk/compare/1.10.0-rc.0...HEAD
+[1.10.0-rc.0]: https://github.com/tokend/new-js-sdk/compare/1.9.0...1.10.0-rc.0
 [1.9.0]: https://github.com/tokend/new-js-sdk/compare/1.9.0-rc.2...1.9.0
 [1.9.0-rc.2]: https://github.com/tokend/new-js-sdk/compare/1.9.0-rc.1...1.9.0-rc.2
 [1.9.0-rc.1]: https://github.com/tokend/new-js-sdk/compare/1.9.0-rc.0...1.9.0-rc.1